review: Atlas-462: clarify folder stuctures and combine malformed file builders
