+++ Atlas.DonorImport.Test/TestHelpers/Models/MalformedDonorFileModels/DonorFileWithoutUpdate.cs
@@ -4,20 +4,13 @@ using System.Text;
 using Atlas.DonorImport.Models.FileSchema;
 using Newtonsoft.Json;
 
-namespace Atlas.DonorImport.Test.TestHelpers.Models
+namespace Atlas.DonorImport.Test.TestHelpers.Models.MalformedDonorFileModels
 {
-    /// <summary>
-    /// Donor import file schema which has a patients column instead of donors.
-    /// </summary>
-    internal class DonorFileWithPatients
+    internal class DonorFileWithoutUpdate
     {
         // ReSharper disable once InconsistentNaming
         [JsonProperty(Order = 1)]
-        public UpdateMode updateMode { get; set; }
-
-        // ReSharper disable once InconsistentNaming
-        [JsonProperty(Order = 2)]
-        public IEnumerable<DonorUpdate> patients { get; set; }
+        public IEnumerable<DonorUpdate> donors { get; set; }
 
         public Stream ToStream()
         {
