+++ Atlas.DonorImport/Exceptions/EmptyDonorFileException.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Atlas.DonorImport.Helpers
+namespace Atlas.DonorImport.Exceptions
 {
     internal class EmptyDonorFileException : Exception
     {
